Copy commands before executing them
Previously when executing a command the command object was executed
directly. This caused problems when it was backed a struct that was not
a pointer, as arguments could not be parsed properly. Furthermore, it
lead to tricky situations during testing, when needing to manually copy
a command.

This commit resolves the issue by two means.

First, commands are now copied before executing them. Second, when a
command is passed as a plain struct, it is automatically lifted to be a
pointer during the cloning process.