stream: Add `ReadLine` and friends
This commit adds three new methods to the stream package called
`ReadLine`, `ReadPassword` and `ReadPasswordStrict`.

The `ReadLine` methods reads from a reader until the end of the line.
The `ReadPassword` and `ReadPasswordStrict` read until the end of the
current line without echoing to the terminal. `ReadPassword` falls back
to `ReadLine` when the input is not a terminal, whereas
`ReadPasswordStrict` returns an error.